CentralManagerを経由してperipheralのメソッド呼ぶ部分をファイル分離
